refactor: corrigir estrutura arquitetural da entidade User (TASK_023)
- Movida entidade User de apps/umbrella_trio_lottopar_web para apps/core
- Criado contexto Admin no apps/core com funções do User
- Adicionadas dependências Ecto ao apps/core (ecto, ecto_sql, postgrex)
- Criado Repo no apps/core para conexão com banco de dados
- Atualizado UserController para usar contexto Admin do apps/core
- Removido contexto Admin duplicado do apps/umbrella_trio_lottopar_web
- Adicionada dependência core no mix.exs do apps/umbrella_trio_lottopar_web
- Projeto compila sem erros
- Credo: 0 issues encontrados
- Arquitetura umbrella agora consistente com TASK_000
