+++ apps/core/lib/umbrella_trio_lottopar_web/repo.ex
@@ -1,0 +1,5 @@
+defmodule UmbrellaTrioLottoparWeb.Repo do
+  use Ecto.Repo,
+    otp_app: :core,
+    adapter: Ecto.Adapters.Postgres
+end
