--- apps/umbrella_trio_lottopar_web/lib/umbrella_trio_lottopar_web/admin.ex
@@ -1,200 +1,0 @@
-defmodule UmbrellaTrioLottoparWeb.Admin do
-  @moduledoc """
-  The Admin context.
-  """
-
-  import Ecto.Query, warn: false
-  alias UmbrellaTrioLottoparWeb.Repo
-
-  alias UmbrellaTrioLottoparWeb.Admin.Operator
-
-  @doc """
-  Returns the list of operators.
-
-  ## Examples
-
-      iex> list_operators()
-      [%Operator{}, ...]
-
-  """
-  def list_operators do
-    Repo.all(Operator)
-  end
-
-  @doc """
-  Gets a single operator.
-
-  Raises `Ecto.NoResultsError` if the Operator does not exist.
-
-  ## Examples
-
-      iex> get_operator!(123)
-      %Operator{}
-
-      iex> get_operator!(456)
-      ** (Ecto.NoResultsError)
-
-  """
-  def get_operator!(id), do: Repo.get!(Operator, id)
-
-  @doc """
-  Creates a operator.
-
-  ## Examples
-
-      iex> create_operator(%{field: value})
-      {:ok, %Operator{}}
-
-      iex> create_operator(%{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def create_operator(attrs) do
-    %Operator{}
-    |> Operator.changeset(attrs)
-    |> Repo.insert()
-  end
-
-  @doc """
-  Updates a operator.
-
-  ## Examples
-
-      iex> update_operator(operator, %{field: new_value})
-      {:ok, %Operator{}}
-
-      iex> update_operator(operator, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def update_operator(%Operator{} = operator, attrs) do
-    operator
-    |> Operator.changeset(attrs)
-    |> Repo.update()
-  end
-
-  @doc """
-  Deletes a operator.
-
-  ## Examples
-
-      iex> delete_operator(operator)
-      {:ok, %Operator{}}
-
-      iex> delete_operator(operator)
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def delete_operator(%Operator{} = operator) do
-    Repo.delete(operator)
-  end
-
-  @doc """
-  Returns an `%Ecto.Changeset{}` for tracking operator changes.
-
-  ## Examples
-
-      iex> change_operator(operator)
-      %Ecto.Changeset{data: %Operator{}}
-
-  """
-  def change_operator(%Operator{} = operator, attrs \\ %{}) do
-    Operator.changeset(operator, attrs)
-  end
-
-  alias UmbrellaTrioLottoparWeb.Admin.User
-
-  @doc """
-  Returns the list of users.
-
-  ## Examples
-
-      iex> list_users()
-      [%User{}, ...]
-
-  """
-  def list_users do
-    Repo.all(User)
-  end
-
-  @doc """
-  Gets a single user.
-
-  Raises `Ecto.NoResultsError` if the User does not exist.
-
-  ## Examples
-
-      iex> get_user!(123)
-      %User{}
-
-      iex> get_user!(456)
-      ** (Ecto.NoResultsError)
-
-  """
-  def get_user!(id), do: Repo.get!(User, id)
-
-  @doc """
-  Creates a user.
-
-  ## Examples
-
-      iex> create_user(%{field: value})
-      {:ok, %User{}}
-
-      iex> create_user(%{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def create_user(attrs) do
-    %User{}
-    |> User.changeset(attrs)
-    |> Repo.insert()
-  end
-
-  @doc """
-  Updates a user.
-
-  ## Examples
-
-      iex> update_user(user, %{field: new_value})
-      {:ok, %User{}}
-
-      iex> update_user(user, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def update_user(%User{} = user, attrs) do
-    user
-    |> User.changeset(attrs)
-    |> Repo.update()
-  end
-
-  @doc """
-  Deletes a user.
-
-  ## Examples
-
-      iex> delete_user(user)
-      {:ok, %User{}}
-
-      iex> delete_user(user)
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def delete_user(%User{} = user) do
-    Repo.delete(user)
-  end
-
-  @doc """
-  Returns an `%Ecto.Changeset{}` for tracking user changes.
-
-  ## Examples
-
-      iex> change_user(user)
-      %Ecto.Changeset{data: %User{}}
-
-  """
-  def change_user(%User{} = user, attrs \\ %{}) do
-    User.changeset(user, attrs)
-  end
-end
