+++ apps/umbrella_trio_lottopar_web/lib/umbrella_trio_lottopar_web_web/controllers/user_controller.ex
@@ -1,25 +1,25 @@
 defmodule UmbrellaTrioLottoparWeb.UserController do
   use UmbrellaTrioLottoparWeb, :controller
 
-  alias UmbrellaTrioLottoparWeb.Admin
+  alias UmbrellaTrioLottoparWeb.Admin, as: CoreAdmin
   alias UmbrellaTrioLottoparWeb.Admin.User
 
   def index(conn, _params) do
-    users = Admin.list_users()
+    users = CoreAdmin.list_users()
     render(conn, :index, users: users)
   end
 
   def new(conn, _params) do
-    changeset = Admin.change_user(%User{})
+    changeset = CoreAdmin.change_user(%User{})
     render(conn, :new, changeset: changeset)
   end
 
   def create(conn, %{"user" => user_params}) do
-    case Admin.create_user(user_params) do
+    case CoreAdmin.create_user(user_params) do
       {:ok, user} ->
         conn
         |> put_flash(:info, "User created successfully.")
-        |> redirect(to: ~p"/users/#{user}")
+        |> redirect(to: ~p"/admin/users/#{user}")
 
       {:error, %Ecto.Changeset{} = changeset} ->
         render(conn, :new, changeset: changeset)
@@ -27,24 +27,24 @@ defmodule UmbrellaTrioLottoparWeb.UserController do
   end
 
   def show(conn, %{"id" => id}) do
-    user = Admin.get_user!(id)
+    user = CoreAdmin.get_user!(id)
     render(conn, :show, user: user)
   end
 
   def edit(conn, %{"id" => id}) do
-    user = Admin.get_user!(id)
-    changeset = Admin.change_user(user)
+    user = CoreAdmin.get_user!(id)
+    changeset = CoreAdmin.change_user(user)
     render(conn, :edit, user: user, changeset: changeset)
   end
 
   def update(conn, %{"id" => id, "user" => user_params}) do
-    user = Admin.get_user!(id)
+    user = CoreAdmin.get_user!(id)
 
-    case Admin.update_user(user, user_params) do
+    case CoreAdmin.update_user(user, user_params) do
       {:ok, user} ->
         conn
         |> put_flash(:info, "User updated successfully.")
-        |> redirect(to: ~p"/users/#{user}")
+        |> redirect(to: ~p"/admin/users/#{user}")
 
       {:error, %Ecto.Changeset{} = changeset} ->
         render(conn, :edit, user: user, changeset: changeset)
@@ -52,11 +52,11 @@ defmodule UmbrellaTrioLottoparWeb.UserController do
   end
 
   def delete(conn, %{"id" => id}) do
-    user = Admin.get_user!(id)
-    {:ok, _user} = Admin.delete_user(user)
+    user = CoreAdmin.get_user!(id)
+    {:ok, _user} = CoreAdmin.delete_user(user)
 
     conn
     |> put_flash(:info, "User deleted successfully.")
-    |> redirect(to: ~p"/users")
+    |> redirect(to: ~p"/admin/users")
   end
 end
